Cover issues caused by merge with main
